# docs/openapi.yaml
openapi: 3.1.0
info:
  title: RdMarket Intelligence API
  version: 1.0.0
  description: Analytical USD/IDR market monitoring, forecasting, and backtesting API.
servers:
  - url: /api/v1
paths:
...
  /economic/indicators:
    get:
      operationId: getEconomicIndicators
      summary: Active economic indicator catalog and latest stored observations
      responses:
        "200":
          description: Indicator catalog with latest values or null when no observations are stored
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseEnvelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          { $ref: "#/components/schemas/EconomicIndicator" }
        "401":
          description: Authentication required when read access is protected
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ResponseEnvelope" }
        "503":
          description: Economic data store unavailable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ResponseEnvelope" }
  /economic/indicators/{code}:
    get:
      operationId: getEconomicIndicator
      summary: Economic indicator metadata and latest stored observation
      parameters:
        - name: code
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Indicator metadata and latest observation
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseEnvelope"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/EconomicIndicator" }
        "401":
          description: Authentication required when read access is protected
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ResponseEnvelope" }
        "404":
          description: Indicator not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ResponseEnvelope" }
        "503":
          description: Economic data store unavailable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ResponseEnvelope" }
  /economic/indicators/{code}/history:
    get:
      operationId: getEconomicIndicatorHistory
      summary: Point-in-time economic observation history and revisions
      parameters:
        - name: code
          in: path
          required: true
          schema: { type: string }
        - name: range
          in: query
          schema: { type: string, enum: [1M, 3M, 1Y, 5Y], default: 1Y }
        - name: start
          in: query
          schema: { type: string, format: date }
        - name: end
          in: query
          schema: { type: string, format: date }
        - name: revisions
          in: query
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: Observations with reference date, release timestamp, retrieval time, and revision
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseEnvelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          { $ref: "#/components/schemas/EconomicObservation" }
        "401":
          description: Authentication required when read access is protected
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ResponseEnvelope" }
        "404":
          description: Indicator not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ResponseEnvelope" }
        "422":
          description: Invalid date range or revisions parameter
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ResponseEnvelope" }
        "503":
          description: Economic data store unavailable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ResponseEnvelope" }
...
  /market/usdidr/history:
    get:
      operationId: getHistory
      summary: Historical observations
      parameters:
        - name: range
          in: query
          schema:
            type: string
            enum: [1D, 7D, 1M, 3M, 6M, 1Y, 5Y]
        - name: start
          in: query
          schema: { type: string, format: date }
        - name: end
          in: query
          schema: { type: string, format: date }
      responses:
        "200":
          description: Historical rate points
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseEnvelope"
        "422":
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseEnvelope"
...
components:
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
    ResponseEnvelope:
      type: object
      required: [data, meta, error]
      properties:
        data: {}
        meta:
          type: object
          properties:
            simulated: { type: boolean }
          additionalProperties: true
        error:
          anyOf:
            - $ref: "#/components/schemas/Error"
            - type: "null"
    EconomicIndicator:
      type: object
      required:
        - code
        - name
        - country
        - category
        - unit
        - frequency
        - seasonal_adjustment
        - source_provider
        - source_series_id
        - source_url
        - license_note
        - description
        - change_mode
        - publication_lag_days
        - latest_value
        - previous_value
        - change
        - reference_date
        - period_end
        - release_timestamp
        - retrieved_at
        - stale
        - staleness_days
        - trend
        - sparkline
      properties:
        code: { type: string }
        name: { type: string }
        country: { type: string }
        category: { type: string }
        unit: { type: string }
        frequency: { type: string, enum: [daily, weekly, monthly, quarterly] }
        seasonal_adjustment: { type: boolean }
        source_provider: { type: string }
        source_series_id: { type: string }
        source_url: { type: string, format: uri }
        license_note: { type: string }
        description: { type: string }
        change_mode: { type: string, enum: [level, percent, bps] }
        publication_lag_days: { type: integer, minimum: 1 }
        latest_value: { type: [string, "null"] }
        previous_value: { type: [string, "null"] }
        change: { type: [string, "null"] }
        reference_date: { type: [string, "null"], format: date-time }
        period_end: { type: [string, "null"], format: date-time }
        release_timestamp: { type: [string, "null"], format: date-time }
        retrieved_at: { type: [string, "null"], format: date-time }
        stale: { type: [boolean, "null"] }
        staleness_days: { type: [integer, "null"] }
        trend: { type: string, enum: [rising, falling, flat, not_enough_data] }
        sparkline:
          type: array
          items: { type: string }
    EconomicObservation:
      type: object
      required:
        [
          reference_date,
          period_end,
          value,
          release_timestamp,
          retrieved_at,
          revision,
          is_latest,
        ]
      properties:
        reference_date: { type: string, format: date-time }
        period_end: { type: [string, "null"], format: date-time }
        value: { type: string }
        release_timestamp: { type: [string, "null"], format: date-time }
        retrieved_at: { type: string, format: date-time }
        revision: { type: integer, minimum: 0 }
        is_latest: { type: boolean }
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: rdm_session